chore: refacto callout
