+++ src/features/parcours/amo/components/steps/CalloutAmoEnAttente.tsx
@@ -1,40 +1,83 @@
 "use client";
 
+import { useEffect, useRef, useState } from "react";
+import { useAuth } from "@/features/auth/client";
+import { useParcours } from "@/features/parcours/core/context/useParcours";
 import { Amo } from "@/features/parcours/amo";
-import { getAmoChoisie } from "@/features/parcours/amo/actions";
-import { useEffect, useState } from "react";
+import { assignAmoAutomatique, getAmoChoisie } from "@/features/parcours/amo/actions";
+import { StatutValidationAmo } from "../../domain/value-objects";
+import { useAmoMode } from "../../hooks";
+import { AmoMode } from "../../domain/value-objects/departements-amo";
 
-export default function CalloutAmoEnAttente() {
+interface CalloutAmoEnAttenteProps {
+  /** Disponible uniquement en modes OBLIGATOIRE / AV_AMO_FUSIONNES pour déclencher
+   *  l'auto-attribution puis recharger le parcours. */
+  refresh?: () => Promise<void>;
+}
+
+/**
+ * Callout "AMO en attente". Couvre 2 callout selon le mode AMO :
+ *  - Mode OBLIGATOIRE / AV_AMO_FUSIONNES) : Affiche ce callout dès le début du parcours, avec un message d'attente et les coordonnées de l'AMO une fois attribué
+ *  - Mode FACULTATIF (après que l'utilisateur a choisi un AMO)
+ */
+export default function CalloutAmoEnAttente({ refresh }: CalloutAmoEnAttenteProps = {}) {
+  const { user } = useAuth();
+  const { statutAmo } = useParcours();
+  const amoMode = useAmoMode();
+
+  const isAutoAttributionMode = amoMode === AmoMode.OBLIGATOIRE || amoMode === AmoMode.AV_AMO_FUSIONNES;
+
   const [amoChoisie, setAmoChoisie] = useState<Amo | null>(null);
   const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
+  const [isAssigning, setIsAssigning] = useState(isAutoAttributionMode && statutAmo === null);
+  const assignTriggeredRef = useRef(false);
 
+  // Auto-attribution silencieuse au montage (modes OBLIGATOIRE / AV_AMO_FUSIONNES uniquement)
   useEffect(() => {
-    async function loadAmoChoisie() {
-      try {
-        const result = await getAmoChoisie();
-        if (result.success) {
-          setAmoChoisie(result.data);
-        } else {
-          setError(result.error || "Erreur inconnue");
+    if (!isAutoAttributionMode) return;
+    if (statutAmo !== null) return;
+    if (assignTriggeredRef.current) return;
+    assignTriggeredRef.current = true;
+
+    setIsAssigning(true);
+    assignAmoAutomatique()
+      .then((result) => {
+        if (!result.success) {
+          setError(result.error || "Impossible d'attribuer un AMO automatiquement");
+          return;
         }
-      } catch (err) {
-        setError("Erreur lors du chargement");
-        console.error(err);
-      } finally {
-        setIsLoading(false);
+        return refresh?.();
+      })
+      .finally(() => {
+        setIsAssigning(false);
+      });
+  }, [isAutoAttributionMode, statutAmo, refresh]);
+
+  // Charge la fiche de l'AMO une fois la validation existante
+  useEffect(() => {
+    if (statutAmo === null) return;
+    getAmoChoisie().then((result) => {
+      if (result.success) {
+        setAmoChoisie(result.data);
       }
-    }
+    });
+  }, [statutAmo]);
 
-    loadAmoChoisie();
-  }, []);
-
-  if (isLoading) {
+  if (isAssigning) {
     return (
       <div className="fr-callout">
-        <p className="fr-callout__text">Chargement des informations AMO...</p>
+        <p className="fr-callout__text">Attribution de votre AMO local en cours...</p>
       </div>
     );
+  }
+
+  // Statuts terminaux : laisse les autres callouts prendre le relais
+  if (
+    statutAmo === StatutValidationAmo.LOGEMENT_ELIGIBLE ||
+    statutAmo === StatutValidationAmo.LOGEMENT_NON_ELIGIBLE ||
+    statutAmo === StatutValidationAmo.ACCOMPAGNEMENT_REFUSE
+  ) {
+    return null;
   }
 
   return (
@@ -47,28 +90,50 @@ export default function CalloutAmoEnAttente() {
       )}
 
       <div className="fr-callout fr-icon-info-line fr-callout--blue-ecume">
-        <p className="fr-callout__title">Nous attendons la réponse de l’AMO</p>
-        <p className="fr-callout__text fr-mb-4w">
-          Votre demande a été envoyée avec succès. Vous recevrez une notification dès que l’AMO aura validé votre
-          collaboration. Si vous ne recevez pas de réponse rapidement, n’hésitez pas à le relancer pour faire avancer
-          votre dossier.
-        </p>
+        {isAutoAttributionMode ? (
+          <>
+            <p className="fr-callout__title">Première étape : être accompagné</p>
+            <p className="fr-callout__text fr-mb-4w">
+              Dans votre département, le recours à un AMO (Assistant à Maîtrise d'Ouvrage) est obligatoire pour
+              bénéficier du Fonds Prévention Argile.
+            </p>
+          </>
+        ) : (
+          <p className="fr-callout__title">Nous attendons la réponse de l'AMO</p>
+        )}
+
+        <p className="fr-text--bold fr-mb-2w">Comment ça marche ?</p>
+        <ul className="fr-mb-4w">
+          <li>Nous avons informé votre AMO local afin qu'il vous accompagne dans votre projet.</li>
+          <li>
+            Dès qu'il aura validé votre dossier, vous recevrez une notification sur votre adresse email
+            {user?.email ? ` ${user.email}` : ""}.
+          </li>
+          <li>Pas de réponse sous 5 jours ? N'hésitez pas à le relancer, voici ses coordonnées :</li>
+        </ul>
 
         {amoChoisie && (
-          <div className="fr-card fr-mt-4w">
-            <div className="fr-card__body">
-              <div className="fr-card__content">
-                <div className="fr-card__desc">
-                  <p className="fr-text--bold fr-mb-1v">{amoChoisie.nom}</p>
-                  {amoChoisie.emails && (
-                    <p className="fr-text--sm fr-mb-1v">Emails : {amoChoisie.emails.split(";").join(", ")}</p>
-                  )}
-                  {amoChoisie.telephone && <p className="fr-text--sm fr-mb-1v">Téléphone : {amoChoisie.telephone}</p>}
-                  {amoChoisie.adresse && <p className="fr-text--sm text-gray-500">{amoChoisie.adresse}</p>}
+          <>
+            <p className="fr-text--bold fr-mb-2w">Votre AMO dédié</p>
+            <div className="fr-card">
+              <div className="fr-card__body">
+                <div className="fr-card__content">
+                  <div className="fr-card__desc">
+                    <p className="fr-text--bold fr-mb-1v">{amoChoisie.nom}</p>
+                    {amoChoisie.emails && (
+                      <p className="fr-text--sm fr-mb-1v">
+                        <a className="fr-link" href={`mailto:${amoChoisie.emails.split(";")[0].trim()}`}>
+                          {amoChoisie.emails.split(";").join(", ")}
+                        </a>
+                      </p>
+                    )}
+                    {amoChoisie.telephone && <p className="fr-text--sm fr-mb-1v">{amoChoisie.telephone}</p>}
+                    {amoChoisie.adresse && <p className="fr-text--sm">{amoChoisie.adresse}</p>}
+                  </div>
                 </div>
               </div>
             </div>
-          </div>
+          </>
         )}
       </div>
     </div>
