+++ src/features/parcours/amo/components/steps/index.ts
@@ -1,6 +1,5 @@
 export { default as CalloutAmoTodo } from "./CalloutAmoTodo";
 export { default as CalloutAmoEnAttente } from "./CalloutAmoEnAttente";
 export { default as CalloutAmoLogementNonEligible } from "./CalloutAmoLogementNonEligible";
-export { default as CalloutAmoObligatoire } from "./CalloutAmoObligatoire";
 export { default as CalloutChoixAccompagnement } from "./CalloutChoixAccompagnement";
 export { default as StepDetailAmo } from "./StepDetailAmo";
