+++ src/features/parcours/core/components/MonCompteClient.tsx
@@ -17,7 +17,6 @@ import { DSStatus } from "../../dossiers-ds/domain";
 import {
   CalloutAmoEnAttente,
   CalloutAmoLogementNonEligible,
-  CalloutAmoObligatoire,
   CalloutAmoTodo,
   CalloutChoixAccompagnement,
   CalloutDiagnosticAccepte,
@@ -278,16 +277,15 @@ function renderChoixAmoCallout(
     return <CalloutAmoLogementNonEligible />;
   }
 
-  // Modes OBLIGATOIRE et AV_AMO_FUSIONNES : auto-attribution + wording dédié.
-  // Le composant gère lui-même l'auto-trigger quand statutAmo est null, et reste affiché
-  // tant que la réponse de l'AMO (LOGEMENT_ELIGIBLE/NON_ELIGIBLE/REFUSE) n'arrive pas.
+  // Modes OBLIGATOIRE et AV_AMO_FUSIONNES : `CalloutAmoEnAttente` gère lui-même
+  // l'auto-attribution silencieuse quand statutAmo est null.
   if (amoMode === AmoMode.OBLIGATOIRE || amoMode === AmoMode.AV_AMO_FUSIONNES) {
     if (
       statutAmo === null ||
       statutAmo === StatutValidationAmo.EN_ATTENTE ||
       statutAmo === StatutValidationAmo.SANS_AMO
     ) {
-      return <CalloutAmoObligatoire refresh={refresh} />;
+      return <CalloutAmoEnAttente refresh={refresh} />;
     }
   }
